Move url encoded handling to the web crate
