+++ crates/brace-web-form/Cargo.toml
@@ -15,11 +15,12 @@ path = "src/lib/lib.rs"
 [dependencies]
 actix-http = "0.2"
 actix-web = "1.0.3"
-bytes = "0.4"
+brace-web = { path = "../brace-web" }
 chrono = { version = "0.4", features = ["serde"] }
-encoding_rs = "0.8"
 failure = "0.1"
 futures = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-serde_qs = "0.5"
+
+[dev-dependencies]
+bytes = "0.4"
