+++ crates/brace-web/src/lib/lib.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod parse;
 pub mod redirect;
 pub mod render;
